+++ docs/api/version-history.md
@@ -19,6 +19,8 @@ keywords: "API, Docker, rcli, REST, documentation"
 
 * `GET /info` now returns an empty string, instead of `<unknown>` for `KernelVersion`
   and `OperatingSystem` if the daemon was unable to obtain this information.
+* `GET /info` now returns information about the product license, if a license
+  has been applied to the daemon.
 
 ## V1.38 API changes